python/appserver/requirements.txt
sqlalchemy==1.4.54
requests==2.28.1
django
django-cors-headers
djangorestframework